move persistent storage into global crates scope
chore:
* change debug log level on trace in instrument
* move configuration into persistent storage crate
