+++ Cargo.toml
@@ -28,7 +28,7 @@ dotenv = { version = "0.15.0" }
 global_utils = { path = "./crates/global_utils" }
 mockall = { version = "0.13.1" }
 ordinals = { version = "0.0.15" }
-persistent_storage = { path = "./btc_indexer/crates/persistent_storage" }
+persistent_storage = { path = "./crates/persistent_storage" }
 reqwest = { version = "0.12.23" }
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = { version = "1.0.143" , features = ["raw_value"]}
