+++ btc_indexer/crates/config_parser/Cargo.toml
@@ -12,3 +12,4 @@ serde_json = { workspace = true }
 strum = { workspace = true }
 thiserror = { workspace = true }
 tracing = { workspace = true }
+global_utils = { workspace = true }
