+++ btc_indexer/crates/config_parser/src/config.rs
@@ -6,10 +6,11 @@ use std::{
 
 use bitcoincore_rpc::{Auth, bitcoin::Network};
 use config::{Config, Environment};
+use global_utils::{env_parser, env_parser::EnvParser};
 use serde::{Deserialize, Serialize};
 use tracing::{debug, instrument};
 
-use crate::{config::env_parser::EnvParser, error::ConfigParserError};
+use crate::error::ConfigParserError;
 
 const CONFIG_FOLDER_NAME: &str = "../../infrastructure/configuration";
 const CARGO_MANIFEST_DIR: &str = "CARGO_MANIFEST_DIR";
@@ -18,8 +19,6 @@ pub const SSH_PRIVATE_KEY_PATH: &str = "SSH_PRIVATE_KEY_PATH";
 pub const DEFAULT_APP_PRODUCTION_CONFIG_NAME: &str = "production";
 const DEFAULT_APP_LOCAL_BASE_FILENAME: &str = "base.toml";
 pub const DEFAULT_APP_LOCAL_CONFIG_NAME: &str = "local";
-pub const POSTGRES_TESTING_URL_ENV_NAME: &str = "DATABASE_URL_TESTING";
-pub const POSTGRES_URL_ENV_NAME: &str = "DATABASE_URL";
 pub const BITCOIN_NETWORK: &str = "BITCOIN_NETWORK";
 pub const BITCOIN_RPC_HOST: &str = "BITCOIN_RPC_HOST";
 pub const BITCOIN_RPC_PORT: &str = "BITCOIN_RPC_PORT";
@@ -52,16 +51,6 @@ pub struct ServerConfig {
 pub struct AppConfig {
     pub http_server_ip: String,
     pub http_server_port: u16,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct PostgresDbTestingCredentials {
-    pub url: String,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct PostgresDbCredentials {
-    pub url: String,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
@@ -136,48 +125,6 @@ impl ConfigVariant {
 
 pub fn get_cargo_manifest_dir() -> String {
     std::env::var(CARGO_MANIFEST_DIR).unwrap()
-}
-
-mod env_parser {
-    use crate::error::ConfigParserError;
-
-    pub trait EnvParser {
-        const ENV_NAME: &'static str;
-        fn obtain_env_value() -> crate::error::Result<String> {
-            obtain_env_value(Self::ENV_NAME)
-        }
-    }
-
-    pub fn obtain_env_value(name: impl AsRef<str>) -> crate::error::Result<String> {
-        std::env::var(name.as_ref()).map_err(|err| ConfigParserError::ConfigEnvParseError {
-            missing_var_name: name.as_ref().to_string(),
-            err,
-        })
-    }
-}
-
-impl EnvParser for PostgresDbTestingCredentials {
-    const ENV_NAME: &'static str = POSTGRES_TESTING_URL_ENV_NAME;
-}
-
-impl EnvParser for PostgresDbCredentials {
-    const ENV_NAME: &'static str = POSTGRES_URL_ENV_NAME;
-}
-
-impl PostgresDbTestingCredentials {
-    pub fn new() -> crate::error::Result<Self> {
-        Ok(Self {
-            url: Self::obtain_env_value()?,
-        })
-    }
-}
-
-impl PostgresDbCredentials {
-    pub fn new() -> crate::error::Result<Self> {
-        Ok(Self {
-            url: Self::obtain_env_value()?,
-        })
-    }
 }
 
 impl BtcRpcCredentials {
