+++ btc_indexer/crates/config_parser/src/error.rs
@@ -2,6 +2,7 @@ use std::{env::VarError, net::AddrParseError, num::ParseIntError};
 
 use bitcoincore_rpc::bitcoin::network::ParseNetworkError;
 use config::ConfigError;
+use global_utils::env_parser::EnvParserError;
 use thiserror::Error;
 
 pub type Result<T> = core::result::Result<T, ConfigParserError>;
@@ -12,8 +13,8 @@ pub enum ConfigParserError {
     SocketConversionError(#[from] AddrParseError),
     #[error("Failed to parse address, error: {0}")]
     ConfigMergingError(#[from] ConfigError),
-    #[error("Failed to parse env variable {missing_var_name}, err: {err}, check if it exists and is valid")]
-    ConfigEnvParseError { missing_var_name: String, err: VarError },
+    #[error(transparent)]
+    ConfigEnvParseError(#[from] EnvParserError),
     #[error("Occurred custom error: {0}")]
     Custom(String),
     #[error("Failed to parse integer, var name: {var_name}, err: {err}")]
