+++ btc_indexer/crates/server/tests/test_wallet_tracking.rs
@@ -6,7 +6,7 @@ mod mocked_tx_tracking {
     use axum_test::TestServer;
     use btc_indexer_internals::indexer::{BtcIndexer, IndexerParams, IndexerParamsWithApi};
     use btc_indexer_server::routes::track_wallet::TrackWalletRequest;
-    use config_parser::config::{BtcRpcCredentials, ConfigVariant, PostgresDbCredentials, ServerConfig};
+    use config_parser::config::{BtcRpcCredentials, ConfigVariant, ServerConfig};
     use global_utils::common_types::UrlWrapped;
     use persistent_storage::init::PostgresRepo;
     use tracing::{info, instrument};
