+++ crates/global_utils/Cargo.toml
@@ -14,3 +14,4 @@ serde_json = { workspace = true }
 bitcoin = { workspace = true }
 sqlx = { workspace = true }
 uuid = { workspace = true}
+thiserror = { workspace = true }
