+++ crates/global_utils/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod common_types;
+pub mod env_parser;
 pub mod logger;
