+++ crates/persistent_storage/fixtures/btc_indexer_work_checkpoints.sql
