+++ crates/persistent_storage/fixtures/user_request_stats.sql
