+++ crates/persistent_storage/src/error.rs
