+++ crates/persistent_storage/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod config;
 pub mod error;
 pub mod init;
 pub mod schemas;
