+++ crates/persistent_storage/src/schemas/common.rs
